unify linux and macos into a single module
